Move power monitor logging (fix #288875)
